# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Asynchronous tenant-aware CSV, JSONL, and XLSX export API. Application error codes: 0 success; 10000-19999 common/input; 20000-29999 authentication/authorization; 30000-39999 business; 50000-59999 infrastructure.",
        "title": "Data Export Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/exports/cancel": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Cancel a queued or running export with optimistic locking",
                "parameters": [
                    {
                        "description": "Job and current version",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.VersionedExportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.ExportJobBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/exports/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Create an asynchronous export job",
                "parameters": [
                    {
                        "description": "Export definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.CreateExportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.ExportMutationBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/exports/datasets/describe": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "export datasets"
                ],
                "summary": "Describe an export dataset available to an application",
                "parameters": [
                    {
                        "description": "Dataset selector",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.DescribeExportDatasetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.ExportDatasetDescriptorBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/exports/datasets/list": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "export datasets"
                ],
                "summary": "List export datasets available to an application",
                "parameters": [
                    {
                        "description": "Application scope and search",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.ListExportDatasetsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.ExportDatasetPageBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/exports/download": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Create a short-lived download URL",
                "parameters": [
                    {
                        "description": "Download request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.DownloadExportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.ExportDownloadBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/exports/get": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Get an export job",
                "parameters": [
                    {
                        "description": "Job selector",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.GetExportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.ExportJobBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/exports/list": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Search export jobs for frontend pages",
                "parameters": [
                    {
                        "description": "Filters and pagination",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.ListExportsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.ExportPageBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/exports/retry": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Retry a failed or canceled export",
                "parameters": [
                    {
                        "description": "Retry request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.RetryExportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/httptransport.ExportMutationBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httptransport.CreateExportRequest": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "dataset_code": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "provider_service": {
                    "type": "string"
                },
                "query": {
                    "type": "object"
                },
                "selected_columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.DescribeExportDatasetRequest": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "dataset_code": {
                    "type": "string"
                },
                "provider_service": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.DownloadExportRequest": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "ttl_seconds": {
                    "type": "integer"
                }
            }
        },
...
        "httptransport.ExportDatasetDescriptorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httptransport.ExportColumnBody"
                    }
                },
                "estimated_rows": {
                    "type": "integer"
                },
                "formats": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "supports_snapshot": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "httptransport.ExportDatasetPageBody": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httptransport.ExportDatasetSummaryBody"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "httptransport.ExportDownloadBody": {
            "type": "object",
            "properties": {
                "checksum": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "httptransport.ExportJobBody": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "bytes_written": {
                    "type": "integer"
                },
                "checksum": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "dataset_code": {
                    "type": "string"
                },
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "object_key": {
                    "type": "string"
                },
                "progress_percent": {
                    "type": "integer"
                },
                "provider_service": {
                    "type": "string"
                },
                "query": {
                    "type": "object"
                },
                "rows_exported": {
                    "type": "integer"
                },
                "selected_columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "httptransport.ExportMutationBody": {
            "type": "object",
            "properties": {
                "duplicate": {
                    "type": "boolean"
                },
                "job": {
                    "$ref": "#/definitions/httptransport.ExportJobBody"
                }
            }
        },
...
        "httptransport.GetExportRequest": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.ListExportDatasetsRequest": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.ListExportsRequest": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "created_from": {
                    "type": "string"
                },
                "created_to": {
                    "type": "string"
                },
                "dataset_code": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
...
        "httptransport.Response": {
            "type": "object",
            "properties": {
                "body": {},
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "httptransport.VersionedExportRequest": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Enter \"Bearer {JWT}\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "PSK": {
            "description": "Enter \"PSK {shared-key}\" for routes configured with PSK authentication.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}